+++ samples/demo-app/app/src/main/java/cash/z/wallet/sdk/demoapp/MainActivity.kt
@@ -13,6 +13,7 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import cash.z.wallet.sdk.demoapp.util.DemoConfig
 
 class MainActivity : AppCompatActivity() {
