+++ src/androidTest/java/cash/z/wallet/sdk/db/IntegrationTest.kt
@@ -4,10 +4,9 @@ import androidx.test.platform.app.InstrumentationRegistry
 import cash.z.wallet.sdk.block.CompactBlockDbStore
 import cash.z.wallet.sdk.block.CompactBlockDownloader
 import cash.z.wallet.sdk.block.CompactBlockProcessor
-import cash.z.wallet.sdk.data.PollingTransactionRepository
-import cash.z.wallet.sdk.data.Synchronizer
-import cash.z.wallet.sdk.data.TroubleshootingTwig
-import cash.z.wallet.sdk.data.Twig
+import cash.z.wallet.sdk.transaction.PollingTransactionRepository
+import cash.z.wallet.sdk.ext.TroubleshootingTwig
+import cash.z.wallet.sdk.ext.Twig
 import cash.z.wallet.sdk.ext.SampleSeedProvider
 import cash.z.wallet.sdk.ext.SampleSpendingKeyProvider
 import cash.z.wallet.sdk.jni.RustBackend
