+++ src/androidTest/java/cash/z/wallet/sdk/util/AddressGeneratorUtil.kt
@@ -1,8 +1,8 @@
 package cash.z.wallet.sdk.util
 
 import androidx.test.platform.app.InstrumentationRegistry
-import cash.z.wallet.sdk.data.TroubleshootingTwig
-import cash.z.wallet.sdk.data.Twig
+import cash.z.wallet.sdk.ext.TroubleshootingTwig
+import cash.z.wallet.sdk.ext.Twig
 import cash.z.wallet.sdk.ext.SampleSeedProvider
 import cash.z.wallet.sdk.jni.RustBackend
 import cash.z.wallet.sdk.secure.Wallet
