+++ src/main/java/cash/z/wallet/sdk/transaction/TransactionManager.kt
@@ -1,4 +1,4 @@
-package cash.z.wallet.sdk.data
+package cash.z.wallet.sdk.transaction
 
 import cash.z.wallet.sdk.service.LightWalletService
