+++ src/test/java/cash/z/wallet/sdk/transaction/MockSynchronizerTest.kt
@@ -1,4 +1,4 @@
-package cash.z.wallet.sdk.data
+package cash.z.wallet.sdk.transaction
 
 import cash.z.wallet.sdk.dao.ClearedTransaction
 import kotlinx.coroutines.*
